- renamed to from_generalized_config
