+++ src/core/generator/from_generalized_config.rs
